+++ filter_plugins/filters.py
@@ -1,0 +1,10 @@
+import os
+
+def path_join(parts):
+    return os.path.join(*parts)
+
+class FilterModule(object):
+    def filters(self):
+        return {
+            'cartridge_path_join': path_join,
+        }